+++ golem-api-grpc/proto/golem/rib/expr.proto
@@ -203,9 +203,8 @@ message ArmPattern {
     ConstructorArmPattern constructor = 3;
     LiteralArmPattern literal = 4;
     TupleConstructorArmPattern tuple_constructor = 5;
-    FlagConstructorArmPattern flag_constructor = 6;
-    ListConstructorArmPattern list_constructor = 7;
-    RecordConstructorArmPattern record_constructor = 8;
+    ListConstructorArmPattern list_constructor = 6;
+    RecordConstructorArmPattern record_constructor = 7;
   }
 }
 
@@ -224,10 +223,6 @@ message ConstructorArmPattern {
 }
 
 message TupleConstructorArmPattern {
-  repeated ArmPattern patterns = 1;
-}
-
-message FlagConstructorArmPattern {
   repeated ArmPattern patterns = 1;
 }
