--- golem-rib/src/interpreter/comprehensive_test/data_types.rs
@@ -1,438 +1,0 @@
-use crate::interpreter::comprehensive_test::test_utils;
-#[cfg(test)]
-use golem_wasm_ast::analysis::*;
-
-// Result
-pub(crate) fn result_of_str_type() -> AnalysedType {
-    AnalysedType::Result(TypeResult {
-        ok: Some(Box::new(AnalysedType::Str(TypeStr))),
-        err: Some(Box::new(AnalysedType::Str(TypeStr))),
-    })
-}
-
-pub(crate) fn result_of_number_type() -> AnalysedType {
-    AnalysedType::Result(TypeResult {
-        ok: Some(Box::new(AnalysedType::U64(TypeU64))),
-        err: Some(Box::new(AnalysedType::U64(TypeU64))),
-    })
-}
-
-pub(crate) fn result_of_option_type() -> AnalysedType {
-    AnalysedType::Result(TypeResult {
-        ok: Some(Box::new(option_of_str_type())),
-        err: Some(Box::new(option_of_str_type())),
-    })
-}
-
-pub(crate) fn result_of_variant_type() -> AnalysedType {
-    AnalysedType::Result(TypeResult {
-        ok: Some(Box::new(variant_type())),
-        err: Some(Box::new(variant_type())),
-    })
-}
-
-pub(crate) fn result_of_enum_type() -> AnalysedType {
-    AnalysedType::Result(TypeResult {
-        ok: Some(Box::new(enum_type())),
-        err: Some(Box::new(enum_type())),
-    })
-}
-
-pub(crate) fn result_of_tuple_type() -> AnalysedType {
-    AnalysedType::Result(TypeResult {
-        ok: Some(Box::new(tuple_type())),
-        err: Some(Box::new(tuple_type())),
-    })
-}
-
-pub(crate) fn result_of_flag_type() -> AnalysedType {
-    AnalysedType::Result(TypeResult {
-        ok: Some(Box::new(flag_type())),
-        err: Some(Box::new(flag_type())),
-    })
-}
-
-pub(crate) fn result_of_record_type() -> AnalysedType {
-    AnalysedType::Result(TypeResult {
-        ok: Some(Box::new(record_type())),
-        err: Some(Box::new(record_type())),
-    })
-}
-
-pub(crate) fn result_of_list_type() -> AnalysedType {
-    AnalysedType::Result(TypeResult {
-        ok: Some(Box::new(list_of_str_type())),
-        err: Some(Box::new(list_of_str_type())),
-    })
-}
-
-// List
-pub(crate) fn list_of_number_type_type() -> AnalysedType {
-    AnalysedType::List(TypeList {
-        inner: Box::new(AnalysedType::U64(TypeU64)),
-    })
-}
-
-pub(crate) fn list_of_str_type() -> AnalysedType {
-    AnalysedType::List(TypeList {
-        inner: Box::new(AnalysedType::Str(TypeStr)),
-    })
-}
-
-pub(crate) fn list_of_option_type() -> AnalysedType {
-    AnalysedType::List(TypeList {
-        inner: Box::new(AnalysedType::Option(TypeOption {
-            inner: Box::new(AnalysedType::Str(TypeStr)),
-        })),
-    })
-}
-
-pub(crate) fn list_of_list_type() -> AnalysedType {
-    AnalysedType::List(TypeList {
-        inner: Box::new(AnalysedType::List(TypeList {
-            inner: Box::new(AnalysedType::Str(TypeStr)),
-        })),
-    })
-}
-
-pub(crate) fn list_of_variant_type() -> AnalysedType {
-    AnalysedType::List(TypeList {
-        inner: Box::new(variant_type()),
-    })
-}
-
-pub(crate) fn list_of_enum_type() -> AnalysedType {
-    AnalysedType::List(TypeList {
-        inner: Box::new(enum_type()),
-    })
-}
-
-pub(crate) fn list_of_tuple() -> AnalysedType {
-    AnalysedType::List(TypeList {
-        inner: Box::new(tuple_type()),
-    })
-}
-
-pub(crate) fn list_of_record_type() -> AnalysedType {
-    AnalysedType::List(TypeList {
-        inner: Box::new(record_type()),
-    })
-}
-
-pub(crate) fn option_of_number_type() -> AnalysedType {
-    AnalysedType::Option(TypeOption {
-        inner: Box::new(AnalysedType::U64(TypeU64)),
-    })
-}
-
-// Option
-pub(crate) fn option_of_str_type() -> AnalysedType {
-    AnalysedType::Option(TypeOption {
-        inner: Box::new(AnalysedType::Str(TypeStr)),
-    })
-}
-
-pub(crate) fn option_of_option_type() -> AnalysedType {
-    AnalysedType::Option(TypeOption {
-        inner: Box::new(AnalysedType::Option(TypeOption {
-            inner: Box::new(AnalysedType::Str(TypeStr)),
-        })),
-    })
-}
-
-pub(crate) fn option_of_variant_type() -> AnalysedType {
-    AnalysedType::Option(TypeOption {
-        inner: Box::new(variant_type()),
-    })
-}
-
-pub(crate) fn option_of_enum_type() -> AnalysedType {
-    AnalysedType::Option(TypeOption {
-        inner: Box::new(enum_type()),
-    })
-}
-
-pub(crate) fn option_of_tuple() -> AnalysedType {
-    AnalysedType::Option(TypeOption {
-        inner: Box::new(tuple_type()),
-    })
-}
-
-pub(crate) fn option_of_record_type() -> AnalysedType {
-    AnalysedType::Option(TypeOption {
-        inner: Box::new(record_type()),
-    })
-}
-
-pub(crate) fn option_of_list() -> AnalysedType {
-    AnalysedType::Option(TypeOption {
-        inner: Box::new(AnalysedType::List(TypeList {
-            inner: Box::new(AnalysedType::Str(TypeStr)),
-        })),
-    })
-}
-
-// Record
-pub(crate) fn record_type() -> AnalysedType {
-    test_utils::analysed_type_record(vec![
-        (
-            "string-headers",
-            test_utils::analysed_type_record(vec![(
-                "authorization-string",
-                AnalysedType::Str(TypeStr),
-            )]),
-        ),
-        (
-            "data-body",
-            test_utils::analysed_type_record(vec![
-                ("str", AnalysedType::Str(TypeStr)),
-                ("list-of-str", list_of_str_type()),
-                ("list-of-option", list_of_option_type()),
-                ("list-of-list", list_of_list_type()),
-                ("list-of-variant", list_of_variant_type()),
-                ("list-of-enum", list_of_enum_type()),
-                ("list-of-tuple", list_of_tuple()),
-                (
-                    "list-of-record",
-                    AnalysedType::List(TypeList {
-                        inner: Box::new(test_utils::analysed_type_record(vec![
-                            ("field-string-one", AnalysedType::Str(TypeStr)),
-                            ("field-string-two", AnalysedType::Str(TypeStr)),
-                        ])),
-                    }),
-                ),
-                ("option-of-str", option_of_str_type()),
-                ("option-of-option", option_of_option_type()),
-                ("option-of-variant", option_of_variant_type()),
-                ("option-of-enum", option_of_enum_type()),
-                ("option-of-tuple", option_of_tuple()),
-                (
-                    "option-of-record",
-                    AnalysedType::Option(TypeOption {
-                        inner: Box::new(test_utils::analysed_type_record(vec![
-                            ("field-string-one", AnalysedType::Str(TypeStr)),
-                            ("field-string-two", AnalysedType::Str(TypeStr)),
-                        ])),
-                    }),
-                ),
-                ("option-of-list", option_of_list()),
-                (
-                    "nested-record",
-                    test_utils::analysed_type_record(vec![
-                        ("field-string-one", AnalysedType::Str(TypeStr)),
-                        ("field-string-two", AnalysedType::Str(TypeStr)),
-                    ]),
-                ),
-                ("variant-data-a", variant_type()),
-                ("variant-data-b", variant_type()),
-                ("variant-data-c", variant_type()),
-                ("variant-data-d", variant_type()),
-                ("variant-data-e", variant_type()),
-                ("variant-data-f", variant_type()),
-                ("enum-data-a", enum_type()),
-                ("enum-data-b", enum_type()),
-                ("enum-data-c", enum_type()),
-                ("flags-data-a", flag_type()),
-                ("flags-data-b", flag_type()),
-                ("flags-data-c", flag_type()),
-                ("result-data-a", result_of_str_type()),
-                ("result-data-b", result_of_number_type()),
-                ("result-data-c", result_of_enum_type()),
-                ("result-data-d", result_of_variant_type()),
-                ("result-data-e", result_of_tuple_type()),
-                ("result-data-f", result_of_option_type()),
-                ("result-data-g", result_of_str_type()),
-                ("result-data-h", result_of_number_type()),
-                ("result-data-i", result_of_enum_type()),
-                ("result-data-j", result_of_variant_type()),
-                ("result-data-k", result_of_tuple_type()),
-                ("result-data-l", result_of_option_type()),
-                ("result-data-m", result_of_flag_type()),
-                ("result-data-n", result_of_flag_type()),
-                (
-                    "tuple-data",
-                    tuple_type(),
-                ),
-                ("character-data", AnalysedType::Chr(TypeChr)),
-                ("f64-data", AnalysedType::F64(TypeF64)),
-                ("f32-data", AnalysedType::F32(TypeF32)),
-                ("u64-data", AnalysedType::U64(TypeU64)),
-                ("s64-data", AnalysedType::S64(TypeS64)),
-                ("u32-data", AnalysedType::U32(TypeU32)),
-                ("s32-data", AnalysedType::S32(TypeS32)),
-                ("u16-data", AnalysedType::U16(TypeU16)),
-                ("s16-data", AnalysedType::S16(TypeS16)),
-                ("u8-data", AnalysedType::U8(TypeU8)),
-                ("s8-data", AnalysedType::S8(TypeS8)),
-                ("boolean-data", AnalysedType::Bool(TypeBool)),
-            ]),
-        ),
-    ])
-}
-
-// Tuple
-pub(crate) fn tuple_type() -> AnalysedType {
-    AnalysedType::Tuple(TypeTuple {
-        items: vec![
-            AnalysedType::Str(TypeStr),
-            AnalysedType::U64(TypeU64),
-            AnalysedType::S32(TypeS32),
-            AnalysedType::F32(TypeF32),
-            AnalysedType::F64(TypeF64),
-            AnalysedType::Bool(TypeBool),
-            AnalysedType::Chr(TypeChr),
-            AnalysedType::Option(TypeOption {
-                inner: Box::new(AnalysedType::S16(TypeS16)),
-            }),
-            AnalysedType::Result(TypeResult {
-                ok: Some(Box::new(AnalysedType::U8(TypeU8))),
-                err: Some(Box::new(AnalysedType::S8(TypeS8))),
-            }),
-            AnalysedType::List(TypeList {
-                inner: Box::new(AnalysedType::Bool(TypeBool)),
-            }),
-            AnalysedType::Variant(TypeVariant {
-                cases: vec![
-                    NameOptionTypePair {
-                        name: "case-hello".to_string(),
-                        typ: Some(AnalysedType::F64(TypeF64)),
-                    },
-                    NameOptionTypePair {
-                        name: "case-none".to_string(),
-                        typ: None,
-                    },
-                ],
-            }),
-            AnalysedType::Record(TypeRecord {
-                // Option<Record>
-                fields: vec![
-                    NameTypePair {
-                        name: "field-one".to_string(),
-                        typ: AnalysedType::Bool(TypeBool),
-                    },
-                    NameTypePair {
-                        name: "field-two".to_string(),
-                        typ: AnalysedType::Str(TypeStr),
-                    },
-                ],
-            }),
-        ],
-    })
-}
-
-// Enum
-pub(crate) fn enum_type() -> AnalysedType {
-    AnalysedType::Enum(TypeEnum {
-        cases: vec![
-            "enum-a".to_string(),
-            "enum-b".to_string(),
-            "enum-c".to_string(),
-        ],
-    })
-}
-
-// Str
-pub(crate) fn str_type() -> AnalysedType {
-    AnalysedType::Str(TypeStr)
-}
-
-// Number
-pub(crate) fn number_type() -> AnalysedType {
-    AnalysedType::U64(TypeU64)
-}
-
-// Flag
-pub(crate) fn flag_type() -> AnalysedType {
-    AnalysedType::Flags(TypeFlags {
-        names: vec![
-            "featurex".to_string(),
-            "featurey".to_string(),
-            "featurez".to_string(),
-        ],
-    })
-}
-
-// Variant
-pub(crate) fn   variant_type() -> AnalysedType {
-    AnalysedType::Variant(TypeVariant {
-        cases: vec![
-            NameOptionTypePair {
-                name: "case-none".to_string(),
-                typ: None,
-            },
-            NameOptionTypePair {
-                name: "case-str".to_string(),
-                typ: Some(AnalysedType::Str(TypeStr)), // Variant case for String
-            },
-            NameOptionTypePair {
-                name: "case-u64".to_string(),
-                typ: Some(AnalysedType::U64(TypeU64)), // Variant case for u64
-            },
-            NameOptionTypePair {
-                name: "case-s32".to_string(),
-                typ: Some(AnalysedType::S32(TypeS32)), // Variant case for i32
-            },
-            NameOptionTypePair {
-                name: "case-f32".to_string(),
-                typ: Some(AnalysedType::F32(TypeF32)), // Variant case for f32
-            },
-            NameOptionTypePair {
-                name: "case-f64".to_string(),
-                typ: Some(AnalysedType::F64(TypeF64)), // Variant case for f64
-            },
-            NameOptionTypePair {
-                name: "case-bool".to_string(),
-                typ: Some(AnalysedType::Bool(TypeBool)), // Variant case for bool
-            },
-            NameOptionTypePair {
-                name: "case-chr".to_string(),
-                typ: Some(AnalysedType::Chr(TypeChr)), // Variant case for char
-            },
-            NameOptionTypePair {
-                name: "case-list".to_string(),
-                typ: Some(AnalysedType::List(TypeList {
-                    // Variant case for List
-                    inner: Box::new(AnalysedType::S16(TypeS16)),
-                })),
-            },
-            NameOptionTypePair {
-                name: "case-option".to_string(),
-                typ: Some(AnalysedType::Option(TypeOption {
-                    // Variant case for Option
-                    inner: Box::new(AnalysedType::U16(TypeU16)),
-                })),
-            },
-            NameOptionTypePair {
-                name: "case-result".to_string(),
-                typ: Some(AnalysedType::Result(TypeResult {
-                    // Variant case for Result
-                    ok: Some(Box::new(AnalysedType::U8(TypeU8))),
-                    err: Some(Box::new(AnalysedType::S8(TypeS8))),
-                })),
-            },
-            NameOptionTypePair {
-                name: "case-record".to_string(),
-                typ: Some(AnalysedType::Record(TypeRecord {
-                    // Variant case for Record
-                    fields: vec![
-                        NameTypePair {
-                            name: "field1".to_string(),
-                            typ: AnalysedType::Str(TypeStr),
-                        },
-                        NameTypePair {
-                            name: "field2".to_string(),
-                            typ: AnalysedType::Bool(TypeBool),
-                        },
-                    ],
-                })),
-            },
-            NameOptionTypePair {
-                name: "case-tuple".to_string(),
-                typ: Some(AnalysedType::Tuple(TypeTuple {
-                    // Variant case for Tuple
-                    items: vec![AnalysedType::F32(TypeF32), AnalysedType::U32(TypeU32)],
-                })),
-            },
-        ],
-    })
-}
