+++ golem-rib/src/type_inference/identifier_inference.rs
@@ -158,11 +158,6 @@ mod internal {
                     collect_all_identifiers(pattern, state)
                 }
             }
-            ArmPattern::FlagConstructor(patterns) => {
-                for pattern in patterns {
-                    collect_all_identifiers(pattern, state)
-                }
-            }
             ArmPattern::Literal(expr) => accumulate_types_of_identifiers(&mut *expr, state),
         }
     }
