+++ golem-rib/src/type_inference/type_pull_up.rs
@@ -236,12 +236,6 @@ mod internal {
                 }
             }
 
-            ArmPattern::FlagConstructor(arm_patterns) => {
-                for arm_pattern in arm_patterns {
-                    pull_up_types_of_arm_pattern(arm_pattern)?;
-                }
-            }
-
             ArmPattern::ListConstructor(arm_patterns) => {
                 for arm_pattern in arm_patterns {
                     pull_up_types_of_arm_pattern(arm_pattern)?;
